refactor: organizing entities
