+++ __tests__/workers/personalizedDigestEmail.ts
@@ -2,9 +2,8 @@ import { expectSuccessfulBackground, saveFixtures } from '../helpers';
 import worker from '../../src/workers/personalizedDigestEmail';
 import { DataSource } from 'typeorm';
 import createOrGetConnection from '../../src/db';
-import { Post, Source, User } from '../../src/entity';
+import { Post, Source, User, UserPersonalizedDigest } from '../../src/entity';
 import { usersFixture } from '../fixture/user';
-import { UserPersonalizedDigest } from '../../src/entity/UserPersonalizedDigest';
 import { postsFixture } from '../fixture/post';
 import { sourcesFixture } from '../fixture/source';
 import { sendEmail } from '../../src/common';
