+++ src/entity/Comment.ts
@@ -1,6 +1,6 @@
 import { Column, Entity, Index, ManyToOne, PrimaryColumn } from 'typeorm';
 import { Post } from './posts';
-import { User } from './User';
+import { User } from './user';
 
 @Entity()
 export class Comment {
