+++ src/entity/CommentMention.ts
@@ -1,5 +1,5 @@
 import { Entity, ManyToOne, PrimaryColumn } from 'typeorm';
-import { User } from './User';
+import { User } from './user';
 import { Comment } from './Comment';
 
 @Entity()
