+++ src/entity/CommentUpvote.ts
@@ -1,6 +1,6 @@
 import { Column, Entity, Index, ManyToOne, PrimaryColumn } from 'typeorm';
 import { Comment } from './Comment';
-import { User } from './User';
+import { User } from './user';
 
 @Entity()
 export class CommentUpvote {
