+++ src/entity/DevCard.ts
@@ -5,7 +5,7 @@ import {
   ManyToOne,
   PrimaryGeneratedColumn,
 } from 'typeorm';
-import { User } from './User';
+import { User } from './user';
 
 @Entity()
 export class DevCard {
