+++ src/entity/index.ts
@@ -1,3 +1,4 @@
+export * from './user';
 export * from './AdvancedSettings';
 export * from './Alerts';
 export * from './Banner';
@@ -20,8 +21,6 @@ export * from './SourceMember';
 export * from './SourceRequest';
 export * from './TagSegment';
 export * from './View';
-export * from './User';
-export * from './UserState';
 export * from './Comment';
 export * from './CommentUpvote';
 export * from './Keyword';
@@ -32,6 +31,4 @@ export * from './Submission';
 export * from './notifications';
 export * from './Feature';
 export * from './Invite';
-export * from './UserAction';
 export * from './ContentImage';
-export * from './UserPost';
