+++ src/entity/posts/PostMention.ts
@@ -1,5 +1,5 @@
 import { Entity, ManyToOne, PrimaryColumn } from 'typeorm';
-import { User } from '../User';
+import { User } from '../user';
 import { Post } from './Post';
 
 @Entity()
